re-org functions
